Web/LL4M4/Challenge/Private/web/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
